# Test/review segmented array