--- a/src/eytzinger_benchmark.zig
+++ b/src/eytzinger_benchmark.zig
@@ -105,7 +105,7 @@ fn run_benchmark(comptime layout: Layout, blob: []u8, random: *std.rand.Random) 
             const target = value_picker[v % value_picker.len];
             const page = &pages[page_index];
             const bounds = Eytzinger.search_values(K, V, V.key_compare, &page.keys, &page.values, target);
-            const hit = bounds[binary_search(K, V, V.key_from_value, V.key_compare, bounds, target)];
+            const hit = bounds[binary_search(K, V, V.key_from_value, V.key_compare, bounds, target, .{})];
 
             assert(hit.key == target);
             if (i % pages.len == 0) v += 1;
@@ -136,7 +136,7 @@ fn run_benchmark(comptime layout: Layout, blob: []u8, random: *std.rand.Random) 
         while (i < layout.searches) : (i += 1) {
             const target = value_picker[v % value_picker.len];
             const page = &pages[page_picker[i % page_picker.len]];
-            const hit = page.values[binary_search(K, V, V.key_from_value, V.key_compare, page.values[0..], target)];
+            const hit = page.values[binary_search(K, V, V.key_from_value, V.key_compare, page.values[0..], target, .{})];
 
             assert(hit.key == target);
             if (i % pages.len == 0) v += 1;
@@ -307,7 +307,7 @@ fn binary_search_keys(
     assert(keys.len == layout.keys_count);
     assert(values.len == layout.values_count);
 
-    const key_index = binary_search(Key, Key, V.key_from_key, compare_keys, keys, key);
+    const key_index = binary_search(Key, Key, V.key_from_key, compare_keys, keys, key, .{});
     const key_stride = layout.values_count / layout.keys_count;
     const high = key_index * key_stride;
     if (key_index < keys.len and keys[key_index] == key) {