--- a/src/lsm/forest_fuzz.zig
+++ b/src/lsm/forest_fuzz.zig
@@ -340,47 +340,8 @@ pub fn generate_fuzz_ops(random: std.rand.Random) ![]const FuzzOp {
 }
 
 pub fn main() !void {
-    var seed: ?u64 = null;
-
-    var args = std.process.args();
-
-    // Discard executable name.
-    allocator.free(try args.next(allocator).?);
-
-    while (args.next(allocator)) |arg_or_err| {
-        const arg = try arg_or_err;
-        defer allocator.free(arg);
-
-        if (std.mem.eql(u8, arg, "--seed")) {
-            const seed_string_or_err = args.next(allocator) orelse
-                std.debug.panic("Expected an argument to --seed", .{});
-            const seed_string = try seed_string_or_err;
-            defer allocator.free(seed_string);
-
-            if (seed != null) {
-                std.debug.panic("Received more than one \"--seed\"", .{});
-            }
-            seed = std.fmt.parseInt(u64, seed_string, 10) catch |err|
-                std.debug.panic("Could not parse \"{}\" as an integer seed: {}", .{ std.zig.fmtEscapes(seed_string), err });
-        } else {
-            // When run with `--test-cmd`, `zig run` also passes the location of the zig binary as an extra arg.
-            // I don't know how to turn this off, so we just skip such args.
-            if (!std.mem.endsWith(u8, arg, "zig")) {
-                std.debug.panic("Unrecognized argument: \"{}\"", .{std.zig.fmtEscapes(arg)});
-            }
-        }
-    }
-
-    if (seed == null) {
-        // If no seed was given, use a random seed instead.
-        var buffer: [@sizeOf(u64)]u8 = undefined;
-        try std.os.getrandom(&buffer);
-        seed = @bitCast(u64, buffer);
-    }
-
-    log.info("Fuzz seed = {}", .{seed.?});
-
-    var rng = std.rand.DefaultPrng.init(seed.?);
+    const fuzz_args = try fuzz.parse_fuzz_args(allocator);
+    var rng = std.rand.DefaultPrng.init(fuzz_args.seed);
 
     const fuzz_ops = try generate_fuzz_ops(rng.random());
     defer allocator.free(fuzz_ops);