--- a/src/lsm/table_immutable.zig
+++ b/src/lsm/table_immutable.zig
@@ -124,6 +124,7 @@ pub fn TableImmutableType(comptime Table: type) type {
                 compare_keys,
                 table.values,
                 key,
+                .{},
             );
             if (result.exact) {
                 const value = &table.values[result.index];