--- a/src/test/fuzz.zig
+++ b/src/test/fuzz.zig
@@ -2,6 +2,9 @@
 
 const std = @import("std");
 const assert = std.debug.assert;
+const mem = std.mem;
+
+const log = std.log.scoped(.fuzz);
 
 /// Returns an integer of type `T` with an exponential distribution of rate `avg`.
 /// Note: If you specify a very high rate then `std.math.maxInt(T)` may be over-represented.
@@ -44,4 +47,62 @@ pub fn random_enum(
         if (choice < 0) return @intToEnum(Enum, enum_field.value);
     }
     unreachable;
+}
+
+pub const FuzzArgs = struct {
+    seed: u64,
+};
+
+/// Parse common command-line arguments to fuzzers:
+///
+///    [--seed u64]
+///        Sets the seed used for the random number generator.
+pub fn parse_fuzz_args(allocator: mem.Allocator) !FuzzArgs {
+    var seed: ?u64 = null;
+
+    var args = std.process.args();
+
+    // Discard executable name.
+    allocator.free(try args.next(allocator).?);
+
+    while (args.next(allocator)) |arg_or_err| {
+        const arg = try arg_or_err;
+        defer allocator.free(arg);
+
+        if (std.mem.eql(u8, arg, "--seed")) {
+            const seed_string_or_err = args.next(allocator) orelse
+                std.debug.panic("Expected an argument to --seed", .{});
+            const seed_string = try seed_string_or_err;
+            defer allocator.free(seed_string);
+
+            if (seed != null) {
+                std.debug.panic("Received more than one \"--seed\"", .{});
+            }
+            seed = std.fmt.parseInt(u64, seed_string, 10) catch |err|
+                std.debug.panic(
+                "Could not parse \"{}\" as an integer seed: {}",
+                .{ std.zig.fmtEscapes(seed_string), err },
+            );
+        } else {
+            // When run with `--test-cmd`,
+            // `zig run` also passes the location of the zig binary as an extra arg.
+            // I don't know how to turn this off, so we just skip such args.
+            if (!std.mem.endsWith(u8, arg, "zig")) {
+                std.debug.panic("Unrecognized argument: \"{}\"", .{std.zig.fmtEscapes(arg)});
+            }
+        }
+    }
+
+    if (seed == null) {
+        // If no seed was given, use a random seed instead.
+        var buffer: [@sizeOf(u64)]u8 = undefined;
+        try std.os.getrandom(&buffer);
+        seed = @bitCast(u64, buffer);
+    }
+
+    log.info("Fuzz seed = {}", .{seed.?});
+
+    return FuzzArgs{
+        .seed = seed.?,
+    };
 }