… src/lsm/segmented_array.zig