* Add tests whose insert pattern let's us reach the worst case layout, and test insert/delete at the end of `array.nodes`.
* Remove some redundant branches and function calls.
* Add some extra asserts (eg for `@ptrCast` which does not check size/alignment).
* Simplify the implementation of `insert`.
* Add a `verify` function which documents and asserts the segmented_array invariants.
* Add a `verify` option to SegmentedArrayType. If true, call `array.verify` as pre- and post- condition for every public function.
* Add segmented_array_fuzz which runs the existing tests with a random seed.
* Move some common fuzzing code into test/fuzz.zig

No perf regressions.

On main (with config.verify = false):

```
[16:35:05] jamie@vessel /home/jamie/tigerbeetle
> zig build-exe src/lsm/segmented_array_benchmark.zig -OReleaseSafe
 --main-pkg-path src && sudo benchmark ./segmented_array_benchmark
cset: --> last message, executed args into cpuset "/user", new pid is: 10403
KeyType=u64 ValueCount=_____33 ValueSize=112B NodeSize=___256B LookupTime=____20ns
KeyType=u64 ValueCount=_____34 ValueSize=112B NodeSize=___256B LookupTime=____19ns
KeyType=u64 ValueCount=___1024 ValueSize=112B NodeSize=___256B LookupTime=____89ns
KeyType=u64 ValueCount=___1024 ValueSize=112B NodeSize=___512B LookupTime=___110ns
KeyType=u64 ValueCount=_____64 ValueSize=112B NodeSize=_16384B LookupTime=____17ns
KeyType=u64 ValueCount=____512 ValueSize=112B NodeSize=_16384B LookupTime=____27ns
KeyType=u64 ValueCount=___4096 ValueSize=112B NodeSize=_16384B LookupTime=___152ns
KeyType=u64 ValueCount=__32768 ValueSize=112B NodeSize=_16384B LookupTime=___263ns
KeyType=u64 ValueCount=_262144 ValueSize=112B NodeSize=_16384B LookupTime=___488ns
KeyType=u64 ValueCount=2097152 ValueSize=112B NodeSize=_16384B LookupTime=___987ns

 Performance counter stats for './segmented_array_benchmark':

                75      context-switches
                 0      cpu-migrations
           197,529      page-faults
    40,968,464,171      cpu_core/cpu-cycles/
   101,745,736,225      cpu_core/instructions/
    28,688,028,403      cpu_core/branches/
       243,094,945      cpu_core/branch-misses/
     1,781,177,231      cpu_core/cache-references/
     1,034,736,623      cpu_core/cache-misses/
    48,805,689,075      cpu_core/bus-cycles/
               175      raw_syscalls:sys_enter

      19.555169880 seconds time elapsed

      19.250530000 seconds user
       0.303992000 seconds sys


[16:36:39] jamie@vessel /home/jamie/tigerbeetle
> zig build-exe src/lsm/segmented_array_benchmark.zig -OReleaseSafe --main-pkg-path src && sudo benchmark ./segmented_array_benchmark
cset: --> last message, executed args into cpuset "/user", new pid is: 10447
KeyType=u64 ValueCount=_____33 ValueSize=112B NodeSize=___256B LookupTime=____20ns
KeyType=u64 ValueCount=_____34 ValueSize=112B NodeSize=___256B LookupTime=____19ns
KeyType=u64 ValueCount=___1024 ValueSize=112B NodeSize=___256B LookupTime=____90ns
KeyType=u64 ValueCount=___1024 ValueSize=112B NodeSize=___512B LookupTime=___110ns
KeyType=u64 ValueCount=_____64 ValueSize=112B NodeSize=_16384B LookupTime=____18ns
KeyType=u64 ValueCount=____512 ValueSize=112B NodeSize=_16384B LookupTime=____27ns
KeyType=u64 ValueCount=___4096 ValueSize=112B NodeSize=_16384B LookupTime=___152ns
KeyType=u64 ValueCount=__32768 ValueSize=112B NodeSize=_16384B LookupTime=___264ns
KeyType=u64 ValueCount=_262144 ValueSize=112B NodeSize=_16384B LookupTime=___488ns
KeyType=u64 ValueCount=2097152 ValueSize=112B NodeSize=_16384B LookupTime=___983ns

 Performance counter stats for './segmented_array_benchmark':

                93      context-switches
                 0      cpu-migrations
           197,529      page-faults
    40,946,693,969      cpu_core/cpu-cycles/
   101,746,258,341      cpu_core/instructions/
    28,688,130,575      cpu_core/branches/
       243,400,687      cpu_core/branch-misses/
     1,816,658,640      cpu_core/cache-references/
     1,021,026,768      cpu_core/cache-misses/
    48,779,455,725      cpu_core/bus-cycles/
               175      raw_syscalls:sys_enter

      19.545839347 seconds time elapsed

      19.240307000 seconds user
       0.303989000 seconds sys
```

On this branch:

```
[16:32:04] jamie@vessel /home/jamie/tigerbeetle
> zig build-exe src/lsm/segmented_array_benchmark.zig -OReleaseSafe --main-pkg-path src && sudo bench
mark ./segmented_array_benchmark
cset: --> last message, executed args into cpuset "/user", new pid is: 10075
KeyType=u64 ValueCount=_____33 ValueSize=112B NodeSize=___256B LookupTime=____20ns
KeyType=u64 ValueCount=_____34 ValueSize=112B NodeSize=___256B LookupTime=____20ns
KeyType=u64 ValueCount=___1024 ValueSize=112B NodeSize=___256B LookupTime=____96ns
KeyType=u64 ValueCount=___1024 ValueSize=112B NodeSize=___512B LookupTime=____74ns
KeyType=u64 ValueCount=_____64 ValueSize=112B NodeSize=_16384B LookupTime=____17ns
KeyType=u64 ValueCount=____512 ValueSize=112B NodeSize=_16384B LookupTime=____27ns
KeyType=u64 ValueCount=___4096 ValueSize=112B NodeSize=_16384B LookupTime=___155ns
KeyType=u64 ValueCount=__32768 ValueSize=112B NodeSize=_16384B LookupTime=___284ns
KeyType=u64 ValueCount=_262144 ValueSize=112B NodeSize=_16384B LookupTime=___502ns
KeyType=u64 ValueCount=2097152 ValueSize=112B NodeSize=_16384B LookupTime=___968ns

 Performance counter stats for './segmented_array_benchmark':

             4,385      context-switches
                 0      cpu-migrations
           197,491      page-faults
    40,547,470,292      cpu_core/cpu-cycles/
   101,786,888,922      cpu_core/instructions/
    28,694,719,721      cpu_core/branches/
       236,789,457      cpu_core/branch-misses/
     1,782,098,265      cpu_core/cache-references/
       991,949,040      cpu_core/cache-misses/
    48,326,923,970      cpu_core/bus-cycles/
               175      raw_syscalls:sys_enter

      19.617357773 seconds time elapsed

      19.025033000 seconds user
       0.342550000 seconds sys


[16:32:26] jamie@vessel /home/jamie/tigerbeetle
> zig build-exe src/lsm/segmented_array_benchmark.zig -OReleaseSafe --main-pkg-path src && sudo benchmark ./segmented_array_benchmark
cset: --> last message, executed args into cpuset "/user", new pid is: 10117
KeyType=u64 ValueCount=_____33 ValueSize=112B NodeSize=___256B LookupTime=____20ns
KeyType=u64 ValueCount=_____34 ValueSize=112B NodeSize=___256B LookupTime=____20ns
KeyType=u64 ValueCount=___1024 ValueSize=112B NodeSize=___256B LookupTime=____96ns
KeyType=u64 ValueCount=___1024 ValueSize=112B NodeSize=___512B LookupTime=____75ns
KeyType=u64 ValueCount=_____64 ValueSize=112B NodeSize=_16384B LookupTime=____17ns
KeyType=u64 ValueCount=____512 ValueSize=112B NodeSize=_16384B LookupTime=____27ns
KeyType=u64 ValueCount=___4096 ValueSize=112B NodeSize=_16384B LookupTime=___154ns
KeyType=u64 ValueCount=__32768 ValueSize=112B NodeSize=_16384B LookupTime=___272ns
KeyType=u64 ValueCount=_262144 ValueSize=112B NodeSize=_16384B LookupTime=___483ns
KeyType=u64 ValueCount=2097152 ValueSize=112B NodeSize=_16384B LookupTime=___961ns

 Performance counter stats for './segmented_array_benchmark':

                67      context-switches
                 0      cpu-migrations
           197,491      page-faults
    40,403,793,828      cpu_core/cpu-cycles/
   101,756,955,706      cpu_core/instructions/
    28,687,686,428      cpu_core/branches/
       237,059,774      cpu_core/branch-misses/
     1,746,541,211      cpu_core/cache-references/
       990,559,408      cpu_core/cache-misses/
    48,133,168,525      cpu_core/bus-cycles/
               175      raw_syscalls:sys_enter

      19.286375016 seconds time elapsed

      18.979326000 seconds user
       0.305973000 seconds sys
```