@sentientwaffle 2022-10-15
As part of this PR can you move the segmented array test into its own `segmented_array_fuzz.zig` file? This test has become extremely slow lately (I'm assuming due to the extra `config.verify` checks in binary search). Or maybe we could set things up so that those binary search assertions run only within the binary search tests/fuzzers — I don't want them slowing down the VOPR either.