@jamii 2022-10-15
We discussed in slack splitting config.verify into config.verify_hot and config.verify_cold, with the later gating expensive invariant checks.