@sentientwaffle 2022-10-15
How about something like:

```
/// the default implementations that any importers of this module get:
usingnamespace Module(.{ .verify = false });

fn Module(comptime Config: struct {
    verify: bool,
}) type {
    return struct {
        fn binary_search()
        fn binary_search_raw()
        ...etc
    }
}

test "..." {
    const binary_search = Module(.{ .verify = true }).binary_search;
}
```

This gives us fine control of which verifications are enabled for which tests.

And this pattern is useful more generally to allow unit or fuzz tests to target a specfic config. (Right now many tests are written with a particular config in mind, and would break if that was changed.)

Fuzzing different configs (e.g. varying `lsm_batch_multiple` or `lsm_growth_factor` or `message_size_max`) could expose new bugs.