@jamii 2022-10-15
Can zls handle that? I haven't tested it, but I'd expect that it struggles to autocomplete anything produced by a comptime function. 

We could do `const config = @import("root").tigerbeetle_config` which is less explicit but limits the comptime-ness to just the config instead of every module that depends on it. It also naturally enforces that all modules are using the same config. But on the other hand doesn't allow testing multiple configs from a single compilation unit.