@sentientwaffle 2022-10-15
We already use comptime functions for generics everywhere anyway, though.