@jamii 2022-10-15
On main:

```
> time zig test src/lsm/segmented_array.zig --test-filter Segmented --main-pkg-path src
All 1 tests passed.

________________________________________________________
Executed in   38.15 secs    fish           external
   usr time   38.11 secs    0.00 micros   38.11 secs
   sys time    0.14 secs  701.00 micros    0.14 secs
```

On this branch:

```
> time zig test src/lsm/segmented_array.zig --test-filter Segmented --main-pkg-path src
All 1 tests passed.

________________________________________________________
Executed in   38.56 secs    fish           external
   usr time   38.53 secs  437.00 micros   38.53 secs
   sys time    0.12 secs  135.00 micros    0.12 secs

> time zig build lsm_segmented_array_fuzz
info(fuzz): Fuzz seed = 2374367770311998976

________________________________________________________
Executed in   48.81 secs    fish           external
   usr time   48.26 secs    0.00 micros   48.26 secs
   sys time    0.72 secs  654.00 micros    0.72 secs
```