@sentientwaffle 2022-10-16
Would the segmented array search be sped up significantly if we used a pattern similar to `Options` in the segmented array to disable binary search's verification within the segmented array unit tests? 38 seconds is still really slow for just this one set of tests :confused:.