@jamii 2022-10-17
Oh, I didn't realize it was my changes to binary_search that were making the test slow.

Do we really want to thread these verify flags through all the code as opposed to having a single config.really_verify flag?