@jamii 2022-10-17
I guess we do need a way to write tests over multiple configs anyway...