@sentientwaffle 2022-10-17
We do need a way to test multiple configs, yes. There are a lot of different approaches we could take, and I'm open to suggestions!

It also depends on how much configurability we want/need. e.g. in https://github.com/tigerbeetledb/tigerbeetle/pull/196#issuecomment-1279811158 we don't actually need to thread the parameter everywhere, as long as most callers don't need to modify the configuration. But there are other use cases (e.g. tuning the LSM tree parameters) where we _do_ need to thread the configuration through. For that, I think it would make sense to just have a single "LSM config" object that they all use with a common set of parameters, so that threading it is simple.