@jamii 2022-10-17
```
> time zig test src/lsm/segmented_array.zig --test-filter Segmented
 --main-pkg-path src
All 1 tests passed.

________________________________________________________
Executed in    6.20 secs    fish           external
   usr time    6.17 secs    0.00 micros    6.17 secs
   sys time    0.12 secs  915.00 micros    0.12 secs
```

Compared to a random commit back in Sep:

```
> git status
HEAD detached at 156b047
...

[17:44:01] jamie@vessel /home/jamie/tigerbeetle
> time zig test src/lsm/segmented_array.zig --test-filter Segmented
 --main-pkg-path src
All 1 tests passed.

________________________________________________________
Executed in    5.22 secs    fish           external
   usr time    5.18 secs    0.00 micros    5.18 secs
   sys time    0.14 secs  922.00 micros    0.14 secs
```