@jamii 2022-10-19
Hmm, my proof for node_count_max is not correct - the last node may be less than half-full even when we're splitting earlier nodes. Which suggests a workload that might reach node_count_max_naive...